Upgrade to new yocto rootfs and new debian rootfs
New yocto rootfs gets us Jetpack 6, which (with 6.2 coming soon) will
get us the Nano Super GPU speeds.

This quickly necesitated upgrading tools on x86 to match.  Which is best
done by pointing things to the new sysroot.  The best way to do that is
to refactor the BUILD file generator and share it across the 2
platforms, rather than hand-update anything.

Also, stop linking everything against libc.  It is a bit aggressive and
breaks malloc hooks.

Change-Id: Ia3e8f73870c67b79d340a186b6ded545e718eefa
Signed-off-by: Austin Schuh <austin.linux@gmail.com>