Rename FRC specific code to //frc
Rather than fix some old, dead code, remove some of it.  Some URLs also
got updated to point to new locations.  It is possible to split some of
this up a little bit more if that would be helpful, but mostly this is
just a slog.

A lot of the markdown files are out of date, or no longer helpful.  That
will need another pass.

Change-Id: I9f4dd525803639b15ce870c8c420de6aeb5aedb3
Signed-off-by: Austin Schuh <austin.linux@gmail.com>