Add turbojpeg decoder application
This application subscribes to a CameraImage message with JPEG data
inside it, and decompresses it into a grayscale CameraImage which it
publishes in another channel.

It doesn't drop messages if it gets too far behind.  Benchmarking
suggests that it uses at most 50% of one core at 100 fps, which is
plenty fast enough.

Written with help from Gemini.

Change-Id: I3588d19a05f20d4b6dc32db91139f83078e8aa52
Signed-off-by: Austin Schuh <austin.linux@gmail.com>