Update the foxglove ws-protocol library
I was having trouble with timestamps in the live version of foxglove.
I.e. no matter what I did, I couldn't figure out how to access the
timestamp that we send along with each message. After talking to the
foxglove devs, it turns out that that timestamp is completely ignored
by the client. They have since fixed this issue, but one of the things
I tried to fix it was to upgrade the library.

Since there should be no harm in upgrading the library, I took the
liberty to do so here. This new version should also enable us to send
messages from the client back to the server.

Change-Id: I4f8808100dc671055c64c8a268ced89670d93de8
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>