aos: Add const version of NodeEventLoopFactory::boot_uuid()
I want to expose the boot UUIDs in the mcap files for foxglove. Since
that code currently only has access to a `const NodeEventLoopFactory`,
I need a const version of the `boot_uuid` function. This patch adds
such a version.

It's a bit awkward that the const version has a different
implementation than the non-const version. It was the best option I
could think of.

Change-Id: Iba0d8118d6b43231e012218e0ec4e27ba0b2d43f
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>