Remove CHECK_NOTNULL in preparation for absl upgrade
gtest, ceres, and various other libraries are moving to use absl logging
instead of using glog/gflags.  Because of that, AOS has upgraded to
using ABSL for flags and logging.  In order to update to newer versions
of AOS, we need to stop using APIs which won't be there.

Some of the CHECKs weren't possible to hit because a function will never
return nullptr.  Some were quite unlikely to be hit, a pain to replicate,
and not consistent with the rest of the file, so they got removed.

For at least one of the CHECKs, make AOS more robust to null ptrs to
move a check donw the stack.

Change-Id: I6b934608b862d95b6e198e5a54020ef947cef12a
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>