Make some log reading failures non-fatal errors
This makes it so that a particular error we encountered in a couple of
logs during ingest in the past month or two can be handled non-fatally,
as demonstrated by a log-replay test written against one of those
logs (unfortunately, given that said log contains BRT-specific data, the
test is located in brt/vpu/datalog instead of in the AOS subtree).

This ends up piping through Result<> return values through a significant
portion of the event loop & log reading stacks. Converting further
LOG(FATAL)'s to Result<> errors should be a bit more straightforward now
that the piping is already there.

Note that this doesn't completely handle every possible path that an
error could propagate through---e.g., calls monotonic_now() can trigger
log replay code. Rather than attempt to modify every callsite to
monotonic_now(), I introduced a CheckExpected() inside of
monotonic_now() and if we ever discover a need to make that non-fatal,
we can revisit this decision. The correct solution in that case may involve
altering the log-reading code to avoid the possibility that such a call
could even error.

Additional Notes:
* I have done my best to manually verify that anything which was
  returning a bool or an object which could implicitly transform to a
  bool (smart & raw pointers) were correctly handled.
* To/FromDistributedClock now return Result<>'s, which force any users
  to actually check that the call succeeded for themselves.

Change-Id: Iaab135a3a8bb3924c524cbe59a5db0be93153760
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>