Pull libusb, ssh, rsync, scp and gtk from the sysroot
libusb.so is a linker script.  If you try linking it directly, it ends
up breaking out of the sandbox.  Instead, rely on Bazel to recreate the
effect of it.

Use ssh/rsync/scp as well for deploying code.  One less weird package.

While we are here, gtk is already inside the sysroot.  Use it directly.

Change-Id: I82981356c128f43e67b677399f9e23ce68fc1ab2
Signed-off-by: Austin Schuh <austin.linux@gmail.com>