Pull libusb, ssh, rsync, scp and gtk from the sysroot
libusb.so is a linker script.  If you try linking it directly, it ends
up breaking out of the sandbox.  Instead, rely on Bazel to recreate the
effect of it.

Use ssh/rsync/scp as well for deploying code.  One less weird package.

While we are here, gtk is already inside the sysroot.  Use it directly.

Change-Id: I964880343462d8762b3e15664f4ca27fc0b8c4b7
Signed-off-by: Austin Schuh <austin.linux@gmail.com>