Make message_bridge_test more robust and debuggable
I was getting a test flake that was pretty repeatable.  Digging in, it
only worked when message_bridge would start up on both sides quick
enough to make sure that the queue indices matched across the two nodes.

In reality, we were testing the wrong thing and mixing up queue indices
across the two sides.  The remote queue index as reported by the client
should match the original queue index, not the new queue index.

While we are here, I added a bunch of debugging.  Since it is all VLOGs,
let's check it in.

Change-Id: I2f5e503e01e52f1ef7b9c5f851523e5cf24c055b
Signed-off-by: Austin Schuh <austin.linux@gmail.com>