aos: Use libunwind for stack traces instead of absl
Abseil is having trouble giving useful stack traces. It seems like
others have the same problem.
https://github.com/abseil/abseil-cpp/discussions/1493

I looked at the unwinding code a little bit, but it was beyond my
current capability to understand.

This came up when investigating a segfault that produced no useful
stack trace. When a program crashes due to a `malloc()` we also get a
useful stacktrace now.

Co-authored-by: Philipp Schrader <philipp.schrader@bluerivertech.com>
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Id69fd3dc5b21d2f704070b8ce8b5bda52b9e19a0