Use ABSL_DCHECK for FLATBUFFERS_ASSERT
The bare `assert` call would hit the realtime malloc checks when it
occurred. Rely on `ABSL_DCHECK` instead.

Note: Ostensibly I could use `FLATBUFFERS_ASSERT_INCLUDE` to accomplish
the same thing, but the main reason to do that would be to reduce the
diff present in the subtree; however, using that `#define` would still
require modifying the `BUILD` files in the subtree.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Ifdab69ce8409ed494bfab2b63e58fd649fd68316