Fix use of uninitialized optional in McapLogger
If you generated an MCAP with zero messages in it (as was triggered by
the addition of the --included_channels flag in the log_to_mcap_test),
we would attempt to write an unitinitialized start time to the log.

Unfortunately, this only got caught by msan and not by the
`std::optional`, because we were using `operator->` on the
`std::optional`, which does no error checking (`.value()` does do
error-checking).

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Ia57a7d68399f8bd404411d1a6f244f19997a56d5