Add additional debug information to message_bridge_server
We were encountering situations where setting stream priorities would
fail with `EINVAL`. From inspecting the kernel source, I don't see a
smoking gun but it does appear that it may be possible for the SCTP
association ID to change asynchronously depending on how the network
stack is behaving, which could plausibly end up causing the
`setsockopt()` to fail with the association ID that it has on hand.

If these log statements show up in future crashes indicating that the
association ID is indeed no longer valid, then we can figure out how to
less-fatally handle the error.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I3b0549081e73545a6e9132f423a94afd04fd03f9