Upgrade rules_python to 1.4
This patch does some major refactoring around the pip-related
infrastructure we have set up.

The most noticeable changes are the patches to rules_python. Now that
we have our GCS mirror behaving like a simple pypi index, we can point
pip directly at it. We no longer have to override the URLs.

As part of this, I switched to using `uv` instead of `pip-compile` to
generate the requirements lock file. It is much, much faster and still
produces the same output. Unfortunately, I also had to remove the
hashes from the requirements lock file. There's a weird interaction
with `uv` where it doesn't populate the hashes from all indexes. Once
we migrate to bzlmod and can switch to the uv lock file, that problem
goes away.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ifa3486d30ab13a7b524b5161048beff9c41c24ff