Add set_channel_network_delay_override
Context:
We need a way to specify the network delay for individual channels
while testing LogReader. There was an issue with how we handle logs
that exceed max_network_delay for some channels on a particular node
while other channels on the same node did not exceed max_network_delay.

Change:
Add functions in
aos/events/simulated_event_loop.h
that are similar to set_network_delay, but allow overriding the value
for
a specific channel.

Change-Id: I2976fb7521e0b29d745cd2b0f1c3e4c6904a3008
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>