Slightly clean up subprocess_test
There was a lot of code-sharing going on between tests and it was making
it irritating to write new tests. This moves the event loop creation
into one spot and then makes it so that all of the tests use a similar
method of timeout detection (all the tests had a timer running that
would kill the test if it ran for too long; previously some tests
would run until they hit the timer and then check that a "success"
flag had gotten set. This makes it so that all of the tests now exit
early when they hit their success condition and then fail if they
hit the exit timeout).

Also shard the test because it has no reason not to be sharded and is
relatively long-running.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I3f4b7e765c5df2c9624f9fb1479baf0d0e934662