Add a last_exit_time to the starter status
The variable was already being recorded but was unused. There have
definitely been times when I would have wanted to use this, and I think
it may be useful for helping to track statistics in the future.

Incidentally makes `top` an optional argument to the PopulateStatus
call, since it is sometimes more heavyweight than we want.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Idcd3e3364d93d9ab2bdff4e36cee088f41528caf