Increase max_network_delay to 4s
Context:
We discovered that the SCTP timout was 3 seconds, which caused
some messages that were delivered originally to be silently
dropped when reading the log, due to max_network_delay default
of 1 second. Increasing the value of max_network_delay to a higher
value allowed us to read logs correctly that had previously crashed
LogReader with a fatal error about missing data on a channel in
the middle of a log.

Change
Set the default value of max_network_delay to 4 seconds.

Change-Id: Ib1a2d1a91d0a55a6faa07ebcc088e3bf67fe41e1
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>