aos: Fix message_bridge_client crash on aborted partial msg
We're seeing this crash when a partial delivery is aborted:

    F1211 07:00:21.253086 5358 sctp_lib.h:92] Check failed: !must_be_returned_to_pool : Did you attempt to destroy a Message without using FreeMessage()?
    *** Check failure stack trace: ***
    @ 0xb6c17750db70 absl::lts_20240116::log_internal::LogMessage::SendToLog()
    @ 0xb6c17750d9d8 absl::lts_20240116::log_internal::LogMessage::Flush()
    @ 0xb6c17750def8 absl::lts_20240116::log_internal::LogMessageFatal::~LogMessageFatal()
    @ 0xb6c17744a63c std::__1::vector<>::__base_destruct_at_end[abi:ue170004]()
    @ 0xb6c17749f65c aos::message_bridge::SctpReadWrite::ProcessNotification()
    @ 0xb6c17749e770 aos::message_bridge::SctpReadWrite::ReadMessage()
    @ 0xb6c17746ba90 aos::message_bridge::SctpClientConnection::MessageReceived()
    @ 0xb6c1774aad24 aos::internal::EPoll::InOutEventData::DoCallbacks()
    @ 0xb6c1774aaa90 aos::internal::EPoll::Poll()
    @ 0xb6c1774aa9fc aos::internal::EPoll::Run()
    @ 0xb6c1774a3d74 aos::ShmEventLoop::Run()
    @ 0xb6c177449f20 aos::message_bridge::Main()
    @ 0xb6c17744a23c main
    @ 0xf1c89543b1f0 (unknown)
    @ 0xf1c89543b2cc __libc_start_main
    @ 0xb6c177449730 _start

Similar to (Fix occasional malloc crash in
message_bridge_client), I cannot reproduce this locally so I'm
not sure to write a test case for this. Worse, I couldn't think of a
good way to trigger that code path even with a couple of hacks.

Attempt to fix the issue by placing what I believe is a missing call
to `FreeMessage`. Similar call sites elsewhere in the code make me
think that this is the right answer. The error message hints at this
as well.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I225ec6d4abdde5c8e6ac24ef802e82880671afe0