Fix `com_github_google_benchmark` compilation with new LLVM
I'm seeing this error:

    external/com_github_google_benchmark/src/mutex.h:79:40: error: mutex 'mut_' is still held at the end of function [-Werror,-Wthread-safety-analysis]
       79 |   void lock() ACQUIRE() { mut_.lock(); }
          |                                        ^
    external/com_github_google_benchmark/src/mutex.h:79:32: note: mutex acquired here
       79 |   void lock() ACQUIRE() { mut_.lock(); }
          |                                ^
    external/com_github_google_benchmark/src/mutex.h:80:34: error: releasing mutex 'mut_' that was not held [-Werror,-Wthread-safety-analysis]
       80 |   void unlock() RELEASE() { mut_.unlock(); }
          |                                  ^

The new warning is great, but not useful for mutex wrappers. It's
possible that there's a way to annotate the wrapper to make clang
happy, but I didn't look into it.

Work around the issue by silencing the warning for the repo.

I also took the opportunity to turn the patch into a proper one that I
can apply with `git am`.

This patch adds a new repository entry because it was being pulled in
transitively via GRPC's dependencies before.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I0153e097f31812d6d9e6207778a0f56ccbe6547c